pipelines/requirements.txt
requests
python-dotenv
shapely
pyproj
psycopg2
openai